Renamed plugin file for feature for consistency.
