+++ src/plugins/mod.rs
@@ -1,4 +1,4 @@
-pub mod developer_helpers;
+pub mod smart_asset_io;
 pub mod levels;
 pub mod main_menu;
 pub mod playable_character;
