+++ src/plugins/smart_asset_io.rs
